futex: don't leak robust_list pointer
This prevents an attacker from determining the robust_list or
compat_robust_list userspace pointer of a process created by executing
a setuid binary. Such an attack could be performed by racing
get_robust_list() with a setuid execution. The impact of this issue is
that
an attacker could theoretically bypass ASLR when attacking setuid
binaries.

Link:
https://lore.kernel.org/linux-fsdevel/1477863998-3298-5-git-send-email-jann@thejh.net/

Signed-off-by: Azeem Shaikh <azeemshaikh38@gmail.com>